requirements-research.txt
# Research Platform — Python Dependencies
# Instalar con: pip install -r requirements-research.txt

# ── Data Layer ─────────────────────────────────────────────────────────
MetaTrader5>=5.0.45          # MT5 Python API (requiere MT5 instalado)
duckdb>=0.10.0               # Base de datos local columnar
pandas>=2.0.0
numpy>=1.26.0
pyarrow>=14.0.0              # Parquet I/O

# ── Research Engine ─────────────────────────────────────────────────────
vectorbt>=0.26.0             # WFA vectorizado + Monte Carlo
statsmodels>=0.14.0          # ADF, ARIMA, HMM regime detection

# ── Portfolio Layer ──────────────────────────────────────────────────────
PyPortfolioOpt>=1.5.5        # HRP, Mean-Variance, CVaR

# ── Persistence ──────────────────────────────────────────────────────────
supabase>=2.3.0              # Supabase Python client
python-dotenv>=1.0.0         # .env loading
httpx>=0.27.0                # HTTP client para Cloudflare Worker

# ── Dashboard ────────────────────────────────────────────────────────────
streamlit>=1.32.0
plotly>=5.20.0

# ── Dev / Testing ────────────────────────────────────────────────────────
pytest>=8.0.0
pytest-cov>=4.1.0
pyyaml>=6.0.1                # Config YAML parsing
click>=8.1.7                 # CLI para run_backtest.py
rich>=13.7.0                 # Output bonito en terminal
